# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "book_service",
        "description": "",
        "contact": {
            "email": "asekenskigo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/books/info": {
            "post": {
                "tags": [
                    "books"
                ],
                "summary": "Get Book Information",
                "description": "Returns book information",
                "operationId": "bookInfo",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookInfoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book information is successfully returned.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookInfoResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/": {
            "get": {
                "tags": [
                    "developers"
                ],
                "summary": "Base action",
                "description": "",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Something went wrong..."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BookInfoRequest": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "createdAt": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "descr": {
                                    "type": "string"
                                },
                                "author": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "BookInfoResource": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date"
                    },
                    "title": {
                        "type": "string"
                    },
                    "descr": {
                        "type": "string"
                    },
                    "author": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}